+++ diktat-rules/src/test/resources/test/paragraph3/long_line/LongLineCommentTest2.kt
@@ -1,0 +1,22 @@
+package test.paragraph3.long_line
+
+/**
+ * This rule checks if there is a backing property for field with property accessors, in case they don't use field keyword
+ */
+class ImplicitBackingPropertyRuleTest(configRules: List<RulesConfig>) : DiktatRule(
+    "implicit-backing-property",
+    configRules,
+    listOf(NO_CORRESPONDING_PROPERTY)) {
+    override fun logic(node: ASTNode) {
+        if (node.elementType == CLASS_BODY) {
+            findAllProperties(node)
+        }
+    }
+
+    private fun validateAccessors(node: ASTNode, propsWithBackSymbol: List<String>) {
+        val accessors = node.findAllDescendantsWithSpecificType(PROPERTY_ACCESSOR).filter { it.hasChildOfType(BLOCK) }  // exclude get with expression body
+
+        accessors.filter { it.hasChildOfType(GET_KEYWORD) }.forEach { handleGetAccessors(it, node, propsWithBackSymbol) }
+        accessors.filter { it.hasChildOfType(SET_KEYWORD) }.forEach { handleSetAccessors(it, node, propsWithBackSymbol) }
+    }
+}
